Time: 203 ms (49.93%), Space: 64.8 MB (69.16%) - LeetHub

+++ 1544-count-good-nodes-in-binary-tree/1544-count-good-nodes-in-binary-tree.cs
@@ -18,7 +18,7 @@ public class Solution {
         Verify(root, root.val);
         return good;
     }
-    public void Verify(TreeNode root, long max)
+    public void Verify(TreeNode root, int max)
     {
         if(root == null)    return;
         if(root.val >= max)
